Canadian Student Token (XCAN)
The Canadian Student Token (known by its ticker XCAN) is a Stablecoin that is pegged Canadian dollar on a 1:1 basis. Every unit of this cryptocurrency in circulation is backed up by 1 Canadian Dollar that is held in reserve, in a mix of cash and short-term Canadian Treasury bonds.

The Stablecoin is designed for a world where cashless transactions are becoming more common. Main use cases for the token would be where finance and education come into play. It also allows students and schools to accept payments in digital assets, and shake up the sector of education.

By adding on to the financial and educational system, The Canadian Student Token ensures a safe form of payment and rewarding system.

Overall, the goal is to provide an effective digital payment while encourage students to be more disciplined in school.